+++ tests/test_mask.py
@@ -14,11 +14,11 @@ from ukis_csmask.utils import reclassify, cohen_kappa_score
 @pytest.mark.parametrize(
     "img, band_order, nodata_value",
     [
-        (np.empty((256, 256, 6), np.float32), ["Blue", "Green", "Red", "NIR", "SWIR1", "SWIR2"], None),
-        (np.empty((256, 256, 8), np.float32), ["Green", "Red", "Blue", "NIR", "SWIR1", "SWIR2", "NIR2", "ETC"], None),
-        (np.empty((256, 256, 6), np.float32), ["Green", "Red", "Blue", "NIR", "SWIR1", "SWIR2"], -666),
+        (np.empty((256, 256, 6), np.float32), ["Blue", "Green", "Red", "NIR", "SWIR16", "SWIR22"], None),
+        (np.empty((256, 256, 8), np.float32), ["Green", "Red", "Blue", "NIR", "SWIR16", "SWIR22", "NIR2", "ETC"], None),
+        (np.empty((256, 256, 6), np.float32), ["Green", "Red", "Blue", "NIR", "SWIR16", "SWIR22"], -666),
         (np.empty((256, 256, 4), np.float32), ["Red", "Green", "Blue", "NIR"], 0),
-        (np.empty((256, 256, 5), np.float32), ["Red", "Green", "Blue", "NIR", "SWIR2"], 0),
+        (np.empty((256, 256, 5), np.float32), ["Red", "Green", "Blue", "NIR", "SWIR22"], 0),
     ],
 )
 def test_csmask_init(img, band_order, nodata_value):
@@ -28,11 +28,11 @@ def test_csmask_init(img, band_order, nodata_value):
 @pytest.mark.parametrize(
     "img, band_order, nodata_value",
     [
-        (3, ["Blue", "Green", "Red", "NIR", "SWIR1", "SWIR2"], None),
-        (np.empty((256, 256), np.float32), ["Blue", "Green", "Red", "NIR", "SWIR1", "SWIR2"], None),
-        (np.empty((256, 256, 3), np.float32), ["Blue", "Green", "Red", "NIR", "SWIR1", "SWIR2"], None),
-        (np.empty((256, 256, 6), np.uint8), ["Blue", "Green", "Red", "NIR", "SWIR1", "SWIR2"], None),
-        (np.empty((256, 256, 6), np.float32), ["Blue", "Green", "Yellow", "NIR", "SWIR1", "SWIR2"], None),
+        (3, ["Blue", "Green", "Red", "NIR", "SWIR16", "SWIR22"], None),
+        (np.empty((256, 256), np.float32), ["Blue", "Green", "Red", "NIR", "SWIR16", "SWIR22"], None),
+        (np.empty((256, 256, 3), np.float32), ["Blue", "Green", "Red", "NIR", "SWIR16", "SWIR22"], None),
+        (np.empty((256, 256, 6), np.uint8), ["Blue", "Green", "Red", "NIR", "SWIR16", "SWIR22"], None),
+        (np.empty((256, 256, 6), np.float32), ["Blue", "Green", "Yellow", "NIR", "SWIR16", "SWIR22"], None),
         (np.empty((256, 256, 6), np.float32), None, None),
     ],
 )
@@ -44,8 +44,8 @@ def test_csmask_init_raises(img, band_order, nodata_value):
 @pytest.mark.parametrize(
     "img, band_order, nodata_value",
     [
-        (np.empty((128, 128, 6), np.float32), ["Blue", "Green", "Red", "NIR", "SWIR1", "SWIR2"], None),
-        (np.empty((64, 64, 6), np.float32), ["Green", "Red", "Blue", "NIR", "SWIR1", "SWIR2"], -666),
+        (np.empty((128, 128, 6), np.float32), ["Blue", "Green", "Red", "NIR", "SWIR16", "SWIR22"], None),
+        (np.empty((64, 64, 6), np.float32), ["Green", "Red", "Blue", "NIR", "SWIR16", "SWIR22"], -666),
     ],
 )
 def test_csmask_init_warns(img, band_order, nodata_value):
@@ -64,7 +64,7 @@ def test_csmask_init_warns(img, band_order, nodata_value):
     ],
 )
 def test_csmask_csm_6band(data):
-    csmask = CSmask(img=data["img"], band_order=["Blue", "Green", "Red", "NIR", "SWIR1", "SWIR2"])
+    csmask = CSmask(img=data["img"], band_order=["Blue", "Green", "Red", "NIR", "SWIR16", "SWIR22"])
     y_pred = csmask.csm
     y_true = reclassify(data["msk"], {"reclass_value_from": [0, 1, 2, 3, 4], "reclass_value_to": [2, 0, 0, 0, 1]})
     y_true = y_true.ravel()
@@ -84,7 +84,7 @@ def test_csmask_csm_6band(data):
     ],
 )
 def test_csmask_valid_6band(data):
-    csmask = CSmask(img=data["img"], band_order=["Blue", "Green", "Red", "NIR", "SWIR1", "SWIR2"])
+    csmask = CSmask(img=data["img"], band_order=["Blue", "Green", "Red", "NIR", "SWIR16", "SWIR22"])
     y_pred = csmask.valid
     y_true = reclassify(data["msk"], {"reclass_value_from": [0, 1, 2, 3, 4], "reclass_value_to": [0, 1, 1, 1, 0]})
     y_true_inverted = ~y_true.astype(bool)